// frontend/Lastnext/app/lib/types.tsx
export interface UserProfile {
  id: string;
  user?: string;
  username: string;
  email: string | null;
  profile_image: string | null;
  positions: string;
  properties: Property[];
  created_at: string;

}
// ...
export interface Room {
  room_id: number;
  name: string;
  room_type: string;
  is_active: boolean;
  created_at: string;
  properties?: (number | string | Property)[];
  property_id?: string | number | null;
}
// ...
export interface Property {
  property_id: string;
  name: string;
  description: string | null;
  created_at: string;
  rooms?: Room[];
  id: string;
  
}